server/requirements.txt
Flask
sentence-transformers
google-cloud-storage
google-auth
openai

